// common/types/academics.ts
export type AcademicProps = {
    course: string;
    institution: string; 
    logo?: string | null;
    location?: string;
    academicYears: string[];
    link?: string | null;
    role?: string;
    location_type?: string; 
    responsibilities?: string[] | null;
    indexAcademic?: number;
  };
  